tests: factor grape helper into sep. file
