+++ test/announce.js
@@ -2,8 +2,12 @@
 
 'use strict'
 
-const { Grape } = require('./../')
 const assert = require('assert')
+const {
+  createGrapes,
+  createTwoGrapes
+} = require('./helper.js')
+
 
 describe('service announce', () => {
   it('should find services', (done) => {
@@ -277,48 +281,4 @@ describe('service announce', () => {
 
 function startAnnouncing (grape, name, port) {
   return setInterval(_ => grape.announce(name, port), 20)
-}
-
-function createGrapes (n, onstart) {
-  const grapes = []
-  let missing = n
-
-  for (let i = 0; i < n; i++) {
-    const grape = new Grape({
-      dht_port: 20001 + i,
-      dht_bootstrap: [ '127.0.0.1:' + (20001 + (i + 1) % 2) ],
-      api_port: 40001 + i,
-      dht_peer_maxAge: 200
-    })
-
-    grape.start(() => {
-      if (--missing) return
-      if (onstart) onstart(grapes, stop)
-    })
-    grapes.push(grape)
-  }
-
-  return grapes
-
-  function stop (done) {
-    loop()
-
-    function loop () {
-      if (!grapes.length) return done()
-      grapes.pop().stop(loop)
-    }
-  }
-}
-
-function createTwoGrapes () {
-  const [grape1, grape2] = createGrapes(2)
-  return {grape1, grape2, stop}
-
-  function stop (done) {
-    grape1.stop(_ => {
-      grape2.stop(_ => {
-        done()
-      })
-    })
-  }
 }
